Refactor Android platform-specific implementation.
