+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: daily_pics
 description: A new Flutter application.
-version: 2.0.2+191002
+version: 2.0.2+191008
 
 environment:
   sdk: ">=2.3.0 <3.0.0"
